Feature/feature fix favorites

+++ Exadel.HEH.Backend.Host/Controllers/FavoritesController.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Exadel.HEH.Backend.BusinessLogic.DTOs.Create;
 using Exadel.HEH.Backend.BusinessLogic.DTOs.Get;
 using Exadel.HEH.Backend.BusinessLogic.Services.Abstract;
+using Exadel.HEH.Backend.BusinessLogic.ValidationServices.Abstract;
 using Exadel.HEH.Backend.Host.Controllers.Abstract;
 using Microsoft.AspNetCore.Mvc;
 
@@ -13,11 +14,13 @@ namespace Exadel.HEH.Backend.Host.Controllers
     public class FavoritesController : BaseController<FavoritesDto>
     {
         private readonly IFavoritesService _favoritesService;
+        private readonly IFavoritesValidationService _validationService;
 
-        public FavoritesController(IFavoritesService favoritesService)
+        public FavoritesController(IFavoritesService favoritesService, IFavoritesValidationService validationService)
             : base(favoritesService)
         {
             _favoritesService = favoritesService;
+            _validationService = validationService;
         }
 
         [HttpPost]
@@ -45,9 +48,16 @@ namespace Exadel.HEH.Backend.Host.Controllers
         }
 
         [HttpDelete("{discountId:guid}")]
-        public Task RemoveAsync(Guid discountId)
+        public async Task<ActionResult> RemoveAsync(Guid discountId)
         {
-            return _favoritesService.RemoveAsync(discountId);
+            if (await _validationService.ValidateDiscountIdIsExist(discountId)
+                && await _validationService.ValidateUserFavoritesIsExist(discountId))
+            {
+                await _favoritesService.RemoveAsync(discountId);
+                return Ok();
+            }
+
+            return NotFound();
         }
     }
 }
